Lists which goals of 1996-2012 and part of 2013 were in extra time or overtime
Has variables for when a goal is scored in extra time (ET) or overtime
(OT) in the case of games in the years 2000-2003.

For example, if a player scored in minute 47 and ETgoals=0, this goal
was scored in the 47th minute of the second half.  If ETgoals=1 for
this goal, then it was scored in the second minute of extra time in the
first half.

If a goal was scored in the 93rd minute of a 2002 game and OTgoals=0,
this goal was scored in the 3rd minute of extra time in the second
half.  If OTgoals=1, then this goal was scored in the 3rd minute of the
overtime period.